+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- New `MyPDO::__construct` params, using array contain database configuration.
+- `MyPDO::getInstance()` return single instance.
+
+### Removed
+- Remove database defineder (`DB_NAME`, `DB_HOST`, `DB_USER`, `DB_PASS`).
 
 ## [0.6.1] - 2022-05-19
 ###  Added